[FEATURE] Adds a setting to define, where the generated files are saved
resolves #74

+++ Classes/Hook/GenerateFilesAfterTcaSave.php
@@ -26,6 +26,7 @@ namespace SGalinski\SgCookieOptin\Hook;
  *  This copyright notice MUST APPEAR in all copies of the script!
  ***************************************************************/
 
+use SGalinski\SgCookieOptin\Service\ExtensionSettingsService;
 use SGalinski\SgCookieOptin\Service\LicensingService;
 use SGalinski\SgCookieOptin\Service\MinificationService;
 use SGalinski\SgCookieOptin\Service\TemplateService;
@@ -50,7 +51,6 @@ use TYPO3\CMS\Frontend\Page\PageRepository;
 class GenerateFilesAfterTcaSave {
 	const TABLE_NAME = 'tx_sgcookieoptin_domain_model_optin';
 
-	const FOLDER_FILEADMIN = 'fileadmin/sg_cookie_optin/';
 	const FOLDER_SITEROOT = 'siteroot-#PID#/';
 
 	const TEMPLATE_JAVA_SCRIPT_PATH = 'typo3conf/ext/sg_cookie_optin/Resources/Public/JavaScript/';
@@ -111,7 +111,12 @@ class GenerateFilesAfterTcaSave {
 			return;
 		}
 
-		$folderName = str_replace('#PID#', $siteRoot, self::FOLDER_FILEADMIN . self::FOLDER_SITEROOT);
+		$folder = ExtensionSettingsService::getSetting(ExtensionSettingsService::SETTING_FOLDER);
+		if (!$folder) {
+			return;
+		}
+
+		$folderName = str_replace('#PID#', $siteRoot, $folder . self::FOLDER_SITEROOT);
 		// First remove the folder with all files and then create it again. So no data artifacts are kept.
 		GeneralUtility::rmdir(PATH_site . $folderName, TRUE);
 		GeneralUtility::mkdir_deep(PATH_site . $folderName);
@@ -234,7 +239,7 @@ class GenerateFilesAfterTcaSave {
 			$this->createJsonFile($folderName, $fullData, $cssData, $loadingScripts, $languageUid);
 		}
 
-		GeneralUtility::fixPermissions(PATH_site . self::FOLDER_FILEADMIN, TRUE);
+		GeneralUtility::fixPermissions(PATH_site . $folder, TRUE);
 
 		// reset the TSFE to it's previous state to not influence remaining code
 		$GLOBALS['TSFE'] = $originalTSFE;
